Powerup improvements
Using Backstab to give a losing player a 5-card push is now considered helping.
Using Remove a Random Card to unbust a player when the dealer is bust is now considered helping.
Using Give a card to give a player a special hand or to give a losing player a 5-card push are now considered helping.
Martyr is now considered helping when the dealer has blackjack or the target was bust.
Using Nothing on a dealer with blackjack now gives help rewards.
Fixed a rare error causing Alt. Clear powerups to clear the hand and not lock in place.